64 bits for Spielwiese Last

+++ factory/cf_factor.cc
@@ -128,17 +128,17 @@ void out_cf(const char *s1,const CanonicalForm &f,const char *s2)
     {
       if (CFFactory::gettype()==GaloisFieldDomain)
       {
-         int a= imm2int (f.getval());
+         long a= imm2int (f.getval());
          if ( a == gf_q )
-           printf ("+%d", a);
-         else  if ( a == 0 )
+           printf ("+%ld", a);
+         else  if ( a == 0L )
            printf ("+1");
-         else  if ( a == 1 )
+         else  if ( a == 1L )
            printf ("+%c",gf_name);
          else
          {
            printf ("+%c",gf_name);
-           printf ("^%d",a);
+           printf ("^%ld",a);
          }
       }
       else
